refactor: move common controller logic to controllers/common package
