+++ internal/controller/cluster/group/group.go
@@ -21,7 +21,6 @@ import (
 
 	"github.com/pkg/errors"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/utils/ptr"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -35,7 +34,8 @@ import (
 
 	userv1alpha1cluster "github.com/statnett/provider-cloudian/apis/cluster/user/v1alpha1"
 	apisv1alpha1cluster "github.com/statnett/provider-cloudian/apis/cluster/v1alpha1"
-	userv1alpha1common "github.com/statnett/provider-cloudian/apis/common/user/v1alpha1"
+	controllercommon "github.com/statnett/provider-cloudian/internal/controller/common"
+	groupcontrollercommon "github.com/statnett/provider-cloudian/internal/controller/common/group"
 	"github.com/statnett/provider-cloudian/internal/sdk/cloudian"
 )
 
@@ -52,15 +52,6 @@ const (
 	errUpdateGroup = "cannot update Group"
 )
 
-var (
-	newCloudianService = func(providerConfig *apisv1alpha1cluster.ProviderConfig, authHeader string) (*cloudian.Client, error) {
-		return cloudian.NewClient(
-			providerConfig.Spec.Endpoint,
-			authHeader,
-		), nil
-	}
-)
-
 // Setup adds a controller that reconciles Group managed resources.
 func Setup(mgr ctrl.Manager, o controller.Options) error {
 	name := managed.ControllerName(userv1alpha1cluster.GroupGroupKind)
@@ -70,7 +61,7 @@ func Setup(mgr ctrl.Manager, o controller.Options) error {
 		managed.WithExternalConnector(&connector{
 			kube:         mgr.GetClient(),
 			usage:        resource.NewLegacyProviderConfigUsageTracker(mgr.GetClient(), &apisv1alpha1cluster.ProviderConfigUsage{}),
-			newServiceFn: newCloudianService}),
+			newServiceFn: controllercommon.NewCloudianService}),
 		managed.WithLogger(o.Logger.WithValues("controller", name)),
 		managed.WithPollInterval(o.PollInterval),
 		//nolint:staticcheck // SA1004 crossplane-runtime still depends on deprecated API
@@ -89,7 +80,7 @@ func Setup(mgr ctrl.Manager, o controller.Options) error {
 type connector struct {
 	kube         client.Client
 	usage        *resource.LegacyProviderConfigUsageTracker
-	newServiceFn func(providerConfig *apisv1alpha1cluster.ProviderConfig, authHeader string) (*cloudian.Client, error)
+	newServiceFn func(providerConfigEndpoint string, authHeader string) (*cloudian.Client, error)
 }
 
 // Connect typically produces an ExternalClient by:
@@ -118,7 +109,7 @@ func (c *connector) Connect(ctx context.Context, mg resource.Managed) (managed.E
 		return nil, errors.Wrap(err, errGetCreds)
 	}
 
-	svc, err := c.newServiceFn(pc, string(authHeader))
+	svc, err := c.newServiceFn(pc.Spec.Endpoint, string(authHeader))
 	if err != nil {
 		return nil, errors.Wrap(err, errNewClient)
 	}
@@ -164,7 +155,7 @@ func (c *external) Observe(ctx context.Context, mg resource.Managed) (managed.Ex
 		// Return false when the external resource exists, but it not up to date
 		// with the desired managed resource state. This lets the managed
 		// resource reconciler know that it needs to call Update.
-		ResourceUpToDate: isUpToDate(meta.GetExternalName(mg), cr.Spec.ForProvider, *observedGroup),
+		ResourceUpToDate: groupcontrollercommon.IsUpToDate(meta.GetExternalName(mg), cr.Spec.ForProvider, *observedGroup),
 
 		// Return any details that may be required to connect to the external
 		// resource. These will be stored as the connection secret.
@@ -180,7 +171,7 @@ func (c *external) Create(ctx context.Context, mg resource.Managed) (managed.Ext
 
 	cr.SetConditions(xpv2.Creating())
 
-	if err := c.cloudianService.CreateGroup(ctx, newCloudianGroup(meta.GetExternalName(mg), cr.Spec.ForProvider)); err != nil {
+	if err := c.cloudianService.CreateGroup(ctx, groupcontrollercommon.NewCloudianGroup(meta.GetExternalName(mg), cr.Spec.ForProvider)); err != nil {
 		return managed.ExternalCreation{}, errors.Wrap(err, errCreateGroup)
 	}
 
@@ -197,7 +188,7 @@ func (c *external) Update(ctx context.Context, mg resource.Managed) (managed.Ext
 		return managed.ExternalUpdate{}, errors.New(errNotGroup)
 	}
 
-	if err := c.cloudianService.UpdateGroup(ctx, newCloudianGroup(meta.GetExternalName(mg), cr.Spec.ForProvider)); err != nil {
+	if err := c.cloudianService.UpdateGroup(ctx, groupcontrollercommon.NewCloudianGroup(meta.GetExternalName(mg), cr.Spec.ForProvider)); err != nil {
 		return managed.ExternalUpdate{}, errors.Wrap(err, errUpdateGroup)
 	}
 
@@ -225,23 +216,4 @@ func (c *external) Delete(ctx context.Context, mg resource.Managed) (managed.Ext
 
 func (c *external) Disconnect(ctx context.Context) error {
 	return nil
-}
-
-func isUpToDate(name string, desired userv1alpha1common.GroupParameters, observed cloudian.Group) bool {
-	return newCloudianGroup(name, desired) == observed
-}
-
-func newCloudianGroup(name string, gp userv1alpha1common.GroupParameters) cloudian.Group {
-	return cloudian.Group{
-		Active:             gp.Active,
-		GroupID:            name,
-		GroupName:          gp.GroupName,
-		LDAPEnabled:        ptr.Deref(gp.LDAPEnabled, false),
-		LDAPGroup:          ptr.Deref(gp.LDAPGroup, ""),
-		LDAPMatchAttribute: ptr.Deref(gp.LDAPMatchAttribute, ""),
-		LDAPSearch:         ptr.Deref(gp.LDAPSearch, ""),
-		LDAPSearchUserBase: ptr.Deref(gp.LDAPSearchUserBase, ""),
-		LDAPServerURL:      ptr.Deref(gp.LDAPServerURL, ""),
-		LDAPUserDNTemplate: ptr.Deref(gp.LDAPUserDNTemplate, ""),
-	}
 }
